Using a button to change what happens
so here when you click on the button now, not only does it change
what’s in our <h1> tag, notice that the input field changes as well.
This is the 2 way data binding we’re talking about, anything that
changes where our variable would be, so too our  model. In other words,
the input field is “bound” to our {{greeting}} inside the h1 tag, when
one changes the other changes instantaneously, in either direction (as
in if h1 changes input changes, if input changes h1 changes)